+++ cli.js
@@ -9,4 +9,4 @@ const [, , pathPackageFrom] = process.argv
 if (!pathPackageFrom)
   throw new Error('Missing argument for the path to the package to install.')
 
-installFrom(pathPackageFrom, process.cwd())
+installFrom(pathPackageFrom)