+++ changelog.md
@@ -9,6 +9,7 @@
 
 ### Minor
 
+- The second `installFrom` function `pathPackageTo` argument now defaults to `process.cwd()`.
 - Setup [GitHub Sponsors funding](https://github.com/sponsors/jaydenseric):
   - Added `.github/funding.yml` to display a sponsor button in GitHub.
   - Added a `package.json` `funding` field to enable npm CLI funding features.